Move `VatSerializer` to CustomFields namespace
